@@ -52,6 +52,14 @@
            <artifactId>spring-ai-chroma-store-spring-boot-starter</artifactId>
        </dependency>
        <dependency>
+           <groupId>org.springframework.ai</groupId>
+           <artifactId>spring-ai-anthropic-spring-boot-starter</artifactId>
+       </dependency>
+       <dependency>
+           <groupId>org.springframework.ai</groupId>
+           <artifactId>spring-ai-bedrock-converse-spring-boot-starter</artifactId>
+       </dependency>
+       <dependency>
            <groupId>org.springframework.boot</groupId>
            <artifactId>spring-boot-starter-data-jpa</artifactId>
        </dependency>
@@ -108,8 +116,8 @@
    </build>
 
    <properties>
-       <spring-boot.version>3.3.5</spring-boot.version>
-       <spring-ai.version>1.0.0-M4</spring-ai.version>
+       <spring-boot.version>3.4.1</spring-boot.version>
+       <spring-ai.version>1.0.0-M5</spring-ai.version>
        <junit-jupiter.version>5.9.0</junit-jupiter.version>
    </properties>
 
