@@ -6,6 +6,6 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 @SpringBootApplication
 public class App {
     public static void main(String[] args) {
-        SpringApplication.run(com.baeldung.app.App.class, args);
+        SpringApplication.run(App.class, args);
     }
 }
